# api/openapi.yaml
openapi: 3.0.3
info:
  title: FIAP X - API de Processamento de Vídeos
  description: API distribuída e escalável para envio, acompanhamento de processamento e download de vídeos compactados em ZIP com frames extraídos.
  version: 1.0.0
servers:
  - url: http://localhost:8080/api/v1
    description: Ambiente Local / Docker Compose
paths:
  /auth/register:
    post:
      summary: Registro de novo usuário
      tags: [Autenticação]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name:
                  type: string
                  example: "Investidor FIAP X"
                email:
                  type: string
                  example: "investidor@fiap.com"
                password:
                  type: string
                  format: password
                  example: "senhaSegura123"
      responses:
        201:
          description: Usuário criado com sucesso
        400:
          description: Dados inválidos
        409:
          description: E-mail já cadastrado

  /auth/login:
    post:
      summary: Login de usuário e emissão de JWT
      tags: [Autenticação]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  example: "investidor@fiap.com"
                password:
                  type: string
                  format: password
                  example: "senhaSegura123"
      responses:
        200:
          description: Autenticado com sucesso, retorna token JWT
        401:
          description: Credenciais inválidas

  /videos/upload:
    post:
      summary: Upload de vídeo para processamento assíncrono
      tags: [Vídeos]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [video]
              properties:
                video:
                  type: string
                  format: binary
                  description: Arquivo de vídeo (mp4, avi, mov, mkv, wmv, flv, webm)
      responses:
        202:
          description: Vídeo aceito e enfileirado com sucesso
        400:
          description: Formato de vídeo não suportado ou arquivo ausente
        401:
          description: Não autenticado

  /videos:
    get:
      summary: Listar status dos vídeos do usuário logado
      tags: [Vídeos]
      security:
        - BearerAuth: []
      responses:
        200:
          description: Lista de vídeos do usuário com status
        401:
          description: Não autenticado

  /videos/{id}:
    get:
      summary: Obter detalhes e status de um vídeo
      tags: [Vídeos]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Detalhes do vídeo
        403:
          description: Acesso negado
        404:
          description: Vídeo não encontrado

  /videos/{id}/download:
    get:
      summary: Download do arquivo .ZIP contendo os frames extraídos
      tags: [Vídeos]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Arquivo ZIP com os frames
          content:
            application/zip:
              schema:
                type: string
                format: binary
        400:
          description: Vídeo ainda não concluído
        403:
          description: Acesso negado
        404:
          description: Arquivo não encontrado

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT